+++ CHANGES.md
@@ -103,6 +103,7 @@ Changes in 10.0
 - SQL use timestamp type: standard & without time zone by default in rosariosis.sql
 - CSS add .accounting-totals, .accounting-staff-payroll-totals, .student-billing-totals classes in Expenses.php, Incomes.php, Salaries.php, StaffPayments.php, StudentFees.php & StudentPayments.php
 - SQL rename KEY (reserved keyword) to SORT_KEY for MySQL compatibility in Search.fnc.php, StudentFieldBreakdown.php, StudentBreakdown.php
+- SQL use GROUP BY instead of DISCTINCT ON for MySQL compatibility in Address.inc.php & EnterEligibility.php
 
 Changes in 9.3.1
 ----------------